03_google_ai/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "google-ai-labs"
version = "0.1.0"
description = "Educational Google Gemini API labs with scripts and notebooks."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.68.0",
    "numpy>=2.4.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
notebooks = [
    "ipykernel>=7.2.0",
    "jupyterlab>=4.5.6",
]

[tool.setuptools]
py-modules = []

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
